HSEARCH-3412 Move field projection tests to a dedicated test class
We're going to add even more tests, and the SearchProjectionIT class is
getting quite large... Let's do like we do for Sorts.
